requirements.txt
# CS2KZ Mapping Tools - Python Dependencies

# GUI Frameworks
imgui[glfw]>=2.0.0

# Graphics and Window Management
PyOpenGL>=3.1.0
glfw>=2.0.0
numpy>=1.24.0

# Image Processing
Pillow>=10.0.0

# VTF Support (optional - for Source engine VTF files)
# vtf2img>=0.1.0

# Process Management
psutil>=5.9.0
